Se corrigió un error que consistía en utilizar un tipo de dato incorrecto para obtener el ID de los metodos de pago, ahora se usa un int en lugar de un byte, se corrigió excepciones que provocaban que el programa completo se cierre, ahora solo se cerrarán solamente las ventanas correspondientes, se cambió de color la fuente de las celdas de garantías en las ventanas para insertar venta y modificar venta.

+++ winVenta_Update.xaml.cs
@@ -592,7 +592,7 @@ namespace sisgesoriadao
                         try
                         {
                             DataRowView d = (DataRowView)dgvMetodosPago.SelectedItem;
-                            byte id = byte.Parse(d.Row.ItemArray[0].ToString());
+                            int id = int.Parse(d.Row.ItemArray[0].ToString());
                             double montoUSD, montoBOB;
                             montoUSD = double.Parse(d.Row.ItemArray[1].ToString());
                             montoBOB = double.Parse(d.Row.ItemArray[2].ToString());
